+++ packages/circuit-ui/components/SideNavigation/components/SecondaryLinks/SecondaryLinks.spec.tsx
@@ -23,6 +23,7 @@ import {
   screen,
   type RenderFn,
 } from '../../../../util/test-utils.js';
+import { CircuitError } from '../../../../util/errors.js';
 
 import { SecondaryLinks, type SecondaryLinksProps } from './SecondaryLinks.js';
 
@@ -34,7 +35,7 @@ describe('SecondaryLinks', () => {
     return renderFn(<SecondaryLinks {...props} />);
   }
 
-  const baseProps = {
+  const baseProps: SecondaryLinksProps = {
     secondaryGroups: [
       {
         secondaryLinks: [
@@ -54,6 +55,7 @@ describe('SecondaryLinks', () => {
             href: '/shop/socks',
             onClick: vi.fn(),
             isActive: true,
+            tier: { variant: 'plus' },
           },
         ],
       },
@@ -97,6 +99,43 @@ describe('SecondaryLinks', () => {
     await userEvent.click(screen.getByRole('link'));
 
     expect(onClick).toHaveBeenCalledTimes(1);
+  });
+
+  it('should show a badge when the badge prop is passed', () => {
+    renderSecondaryLinks(render, baseProps);
+    expect(screen.getByText('New')).toBeVisible();
+  });
+
+  it('should show a tier indicator when the tier prop is passed', () => {
+    renderSecondaryLinks(render, baseProps);
+    expect(screen.getByText('plus')).toBeVisible();
+  });
+
+  it('should throw an error if passed both badge and tier props', () => {
+    const invalidProps: SecondaryLinksProps = {
+      secondaryGroups: [
+        {
+          secondaryLinks: [
+            {
+              label: 'Shirts',
+              href: '/shop/shirts',
+              onClick: vi.fn(),
+              badge: {},
+              tier: { variant: 'plus' },
+            },
+          ],
+        },
+      ],
+    };
+
+    const expectedError = new CircuitError(
+      'SideNavigation',
+      'The `badge` and `tier` props cannot be used simultaneously.',
+    );
+
+    expect(() => renderSecondaryLinks(render, invalidProps)).toThrow(
+      expectedError,
+    );
   });
 
   it('should have no accessibility violations', async () => {