+++ packages/circuit-ui/components/SideNavigation/types.ts
@@ -17,6 +17,7 @@ import type { MouseEvent, KeyboardEvent, AnchorHTMLAttributes } from 'react';
 import type { IconComponentType } from '@sumup-oss/icons';
 
 import type { BadgeProps } from '../Badge/index.js';
+import type { TierIndicatorProps } from '../TierIndicator/TierIndicator.js';
 
 export interface PrimaryLinkProps
   extends AnchorHTMLAttributes<HTMLAnchorElement> {
@@ -106,4 +107,8 @@ export interface SecondaryLinkProps {
    * a new link or to indicate new content.
    */
   badge?: BadgeProps;
+  /**
+   * An optional badge to highlight elements belonging to a specific tier.
+   */
+  tier?: Omit<TierIndicatorProps, 'size'>;
 }