+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sumup-oss/icons
 
+## 5.7.0
+
+### Minor Changes
+
+- [#2991](https://github.com/sumup-oss/circuit-ui/pull/2991) [`fea6e1d`](https://github.com/sumup-oss/circuit-ui/commit/fea6e1d61bfd2611394df5c78c651745278039bd) Thanks [@kamenlitchev](https://github.com/kamenlitchev)! - Added a new `Girocard` payment method icon in size 24.
+
+## 5.6.0
+
+### Minor Changes
+
+- [#2956](https://github.com/sumup-oss/circuit-ui/pull/2956) [`8a7ddc7`](https://github.com/sumup-oss/circuit-ui/commit/8a7ddc76658b28c52d1eb375ebf58f603a206a0e) Thanks [@sirineJ](https://github.com/sirineJ)! - Added icons for the Plus tier indicator.
+
+## 5.5.0
+
+### Minor Changes
+
+- [#2950](https://github.com/sumup-oss/circuit-ui/pull/2950) [`0d9195c`](https://github.com/sumup-oss/circuit-ui/commit/0d9195c6e4c8f2456eee607473bddd1ca6385d1a) Thanks [@kamenlitchev](https://github.com/kamenlitchev)! - Added a new `Girocard` card scheme icon in size 32.
+
 ## 5.4.0
 
 ### Minor Changes